proto codegen (#281)
* part 1

* make codegen work for internal w/ protos

* fix

* code review
